Rename styles/styles.ts to styles/common.ts common for clarity of purpose.
