+++ client/app/styles/index.ts
@@ -1,3 +1,3 @@
 export * from './colors'
-export * from './styles'
+export * from './common'
 export * from './mediaQueries'
